refacotr: 모집중인 로드맵 목록 코드 리팩토링 및 반응형
