+++ src/components/Roadmap/RecruitRoadmap/index.tsx
@@ -1,15 +1,11 @@
 import { useEffect, useState } from 'react';
 import { useRouter } from 'next/router';
-import { useGetRoadmaps } from '@/api/hooks/roadmap';
+import RecruitRoadmapList from '@/components/Roadmap/RecruitRoadmap/RecruitRoadmapList';
 import * as Styled from '@/components/Roadmap/RecruitRoadmap/style';
-import CustomSuspense from '@/components/common/CustomSuspense';
 import Input from '@/components/common/Input';
 import Tab from '@/components/common/Tab';
-import { useIntersectionObserver } from '@/hooks/useInterSectionObserver';
 import { useParamsToUrl } from '@/hooks/useParamsToUrl';
 import useQueryParam from '@/hooks/useQueryParam';
-import GroupCard from '../GroupCard';
-import TilyCard from '../TilyCard';
 
 const RecruitRoadmap = () => {
   const router = useRouter();
@@ -29,16 +25,6 @@ const RecruitRoadmap = () => {
       overlapParamsToUrl({ name });
     }
   };
-
-  const { ref, isVisible } = useIntersectionObserver();
-
-  const { data, isLoading, fetchNextPage, hasNextPage } = useGetRoadmaps(router.query);
-
-  useEffect(() => {
-    if (isVisible && hasNextPage) {
-      fetchNextPage();
-    }
-  }, [isVisible, fetchNextPage, hasNextPage, ref]);
 
   useEffect(() => {
     if (router.isReady) {
@@ -90,31 +76,7 @@ const RecruitRoadmap = () => {
           />
         </form>
       </Styled.Navbar>
-
-      <Styled.RoadmapContainer>
-        <CustomSuspense isLoading={isLoading} fallback={<RoadmapSkeleton />}>
-          {category === 'tily'
-            ? data?.map((roadmap) => <TilyCard key={roadmap.id} roadmap={roadmap} />)
-            : data?.map((roadmap) => <GroupCard key={roadmap.id} roadmap={roadmap} />)}
-        </CustomSuspense>
-      </Styled.RoadmapContainer>
-      <Styled.ObserverInterSectionTarget ref={ref} />
-    </>
-  );
-};
-
-const RoadmapSkeleton = () => {
-  return (
-    <>
-      <Styled.Skeleton />
-      <Styled.Skeleton />
-      <Styled.Skeleton />
-      <Styled.Skeleton />
-      <Styled.Skeleton />
-      <Styled.Skeleton />
-      <Styled.Skeleton />
-      <Styled.Skeleton />
-      <Styled.Skeleton />
+      <RecruitRoadmapList />
     </>
   );
 };
