+++ src/components/Roadmap/RecruitRoadmap/style.ts
@@ -1,6 +1,5 @@
 import { css } from '@emotion/react';
 import styled from '@emotion/styled';
-import SkeletonBox from '@/components/common/Skeleton';
 import type { EmotionTheme } from '@/styles/emotion';
 
 export const Navbar = styled.section`
@@ -12,15 +11,15 @@ export const Navbar = styled.section`
     width: 50%;
   }
 
-  @media (max-width: 760px) {
-    flex-wrap: wrap-reverse;
+  @media ${({ theme }) => theme.mediaQuery.md} {
+    flex-direction: column-reverse;
 
     & > div {
-      width: 100%;
+      width: 90%;
     }
 
     & > form {
-      width: 100%;
+      width: 90%;
     }
   }
 `;
@@ -36,32 +35,4 @@ export const InputContainerStyles = (theme: EmotionTheme) => css`
   @media (max-width: 760px) {
     width: 100%;
   }
-`;
-
-export const RoadmapContainer = styled.section`
-  display: grid;
-  grid-template-columns: repeat(4, 1fr);
-  grid-row-gap: 40px;
-
-  @media (max-width: 1024px) {
-    grid-template-columns: repeat(3, 1fr);
-  }
-
-  @media (max-width: 760px) {
-    grid-template-columns: repeat(2, 1fr);
-  }
-
-  @media (max-width: 550px) {
-    grid-template-columns: repeat(1, 1fr);
-  }
-`;
-
-export const ObserverInterSectionTarget = styled.div`
-  width: 100%;
-  height: 2.5rem;
-`;
-
-export const Skeleton = styled(SkeletonBox)`
-  width: 211px;
-  height: 230px;
 `;
